requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0

# OpenAI SDK
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Validation & Schemas
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Data Processing
numpy>=1.24.0

# HTTP Client (if needed)
requests>=2.31.0

# Microsoft Teams Integration
msal>=1.24.0
beautifulsoup4>=4.12.0

# Testing (Development)
pytest>=7.4.0
pytest-cov>=4.1.0
